requirements.txt
PyAudio==0.2.13
numpy==1.25.0
coloredlogs==15.0.1
scipy==1.11.1
Werkzeug==2.3.6
Flask==2.3.2
Flask_Login==0.6.2
Jinja2==3.1.2
waitress==2.1.2
rpi_ws281x==5.0.0
icmplib==3.0.3
Flask-OpenAPI3-UI==1.6.0
jsonschema==4.17.3
APScheduler==3.10.1
psutil
